Move validation stuff from files module to validation module
