+++ dbt_copilot_helper/commands/bootstrap.py
@@ -13,11 +13,11 @@ from dbt_copilot_helper.utils.aws import get_ssm_secret_names
 from dbt_copilot_helper.utils.aws import get_ssm_secrets
 from dbt_copilot_helper.utils.aws import set_ssm_param
 from dbt_copilot_helper.utils.click import ClickDocOptGroup
-from dbt_copilot_helper.utils.files import BOOTSTRAP_SCHEMA
 from dbt_copilot_helper.utils.files import load_and_validate_config
 from dbt_copilot_helper.utils.files import mkfile
 from dbt_copilot_helper.utils.files import to_yaml
 from dbt_copilot_helper.utils.template import setup_templates
+from dbt_copilot_helper.utils.validation import BOOTSTRAP_SCHEMA
 from dbt_copilot_helper.utils.versioning import (
     check_copilot_helper_version_needs_update,
 )
