+++ dbt_copilot_helper/utils/validation.py
@@ -1,5 +1,8 @@
 import re
 
+from schema import Optional
+from schema import Or
+from schema import Schema
 from schema import SchemaError
 
 
@@ -12,3 +15,98 @@ def validate_string(regex_pattern):
         return string
 
     return validator
+
+
+range_validator = validate_string(r"^\d+-\d+$")
+seconds_validator = validate_string(r"^\d+s$")
+
+BOOTSTRAP_SCHEMA = Schema(
+    {
+        "app": str,
+        "environments": {str: {Optional("certificate_arns"): [str]}},
+        "services": [
+            {
+                "name": str,
+                "type": lambda s: s
+                in (
+                    "public",
+                    "backend",
+                ),
+                "repo": str,
+                "image_location": str,
+                Optional("notes"): str,
+                Optional("secrets_from"): str,
+                "environments": {
+                    str: {
+                        "paas": str,
+                        Optional("url"): str,
+                        Optional("ipfilter"): bool,
+                        Optional("memory"): int,
+                        Optional("count"): Or(
+                            int,
+                            {
+                                # https://aws.github.io/copilot-cli/docs/manifest/lb-web-service/#count
+                                "range": range_validator,  # e.g. 1-10
+                                Optional("cooldown"): {
+                                    "in": seconds_validator,  # e.g 30s
+                                    "out": seconds_validator,  # e.g 30s
+                                },
+                                Optional("cpu_percentage"): int,
+                                Optional("memory_percentage"): Or(
+                                    int,
+                                    {
+                                        "value": int,
+                                        "cooldown": {
+                                            "in": seconds_validator,  # e.g. 80s
+                                            "out": seconds_validator,  # e.g 160s
+                                        },
+                                    },
+                                ),
+                                Optional("requests"): int,
+                                Optional("response_time"): seconds_validator,  # e.g. 2s
+                            },
+                        ),
+                    },
+                },
+                Optional("backing-services"): [
+                    {
+                        "name": str,
+                        "type": lambda s: s
+                        in (
+                            "s3",
+                            "s3-policy",
+                            "aurora-postgres",
+                            "rds-postgres",
+                            "redis",
+                            "opensearch",
+                        ),
+                        Optional("paas-description"): str,
+                        Optional("paas-instance"): str,
+                        Optional("notes"): str,
+                        Optional("bucket_name"): str,  # for external-s3 type
+                        Optional("readonly"): bool,  # for external-s3 type
+                        Optional("shared"): bool,
+                    },
+                ],
+                Optional("overlapping_secrets"): [str],
+                "secrets": {
+                    Optional(str): str,
+                },
+                "env_vars": {
+                    Optional(str): str,
+                },
+            },
+        ],
+    },
+)
+
+PIPELINES_SCHEMA = Schema(
+    {
+        "environments": [
+            {
+                "name": str,
+                Optional("requires_approval"): bool,
+            },
+        ],
+    },
+)
