+++ tests/test_command_bootstrap.py
@@ -20,8 +20,8 @@ from dbt_copilot_helper.commands.bootstrap import get_paas_env_vars
 from dbt_copilot_helper.commands.bootstrap import make_config
 from dbt_copilot_helper.commands.bootstrap import migrate_secrets
 from dbt_copilot_helper.utils.aws import set_ssm_param
-from dbt_copilot_helper.utils.files import BOOTSTRAP_SCHEMA
 from dbt_copilot_helper.utils.files import load_and_validate_config
+from dbt_copilot_helper.utils.validation import BOOTSTRAP_SCHEMA
 from tests.conftest import FIXTURES_DIR
 from tests.conftest import TEST_APP_DIR
