+++ tests/utils/test_files.py
@@ -1,10 +1,10 @@
 import pytest
 import yaml
 
-from dbt_copilot_helper.utils.files import BOOTSTRAP_SCHEMA
-from dbt_copilot_helper.utils.files import PIPELINES_SCHEMA
 from dbt_copilot_helper.utils.files import load_and_validate_config
 from dbt_copilot_helper.utils.files import mkfile
+from dbt_copilot_helper.utils.validation import BOOTSTRAP_SCHEMA
+from dbt_copilot_helper.utils.validation import PIPELINES_SCHEMA
 from tests.conftest import FIXTURES_DIR
